remove Upadting uniform variable directly ( Uniform variable isn't portable. There is no uniform variable concept in dx11. So Only Uniform buffer object ( Constant Buffer ) is used )

+++ DX11GraphicsAPI/DX11GraphicsAPI.cpp
@@ -1416,7 +1416,7 @@ namespace dooms
         (
             const unsigned long long bufferObject,
             const GraphicsAPI::eBufferTarget bindBufferTarget,
-            const unsigned int offset,
+            const unsigned long long offset,
             const unsigned long long dataSize,
             const void* const data
         )
